requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.25.0
scikit-learn>=1.3.0
